Break out experiment mean metric tests
